+++ src/sound/al_Spatializer.cpp
@@ -4,7 +4,7 @@ using namespace  al;
 
 Spatializer::Spatializer(const SpeakerLayout& sl)
 {
-    unsigned numSpeakers = sl.speakers().size();
+    size_t numSpeakers = sl.speakers().size();
     for(unsigned i=0;i<numSpeakers;++i){
         mSpeakers.push_back(sl.speakers()[i]);
     }
